+++ kotlin-native/Interop/Runtime/src/jvm/kotlin/kotlinx/cinterop/JvmUtils.kt
@@ -47,7 +47,7 @@ fun bitsToDouble(bits: Long): Double = java.lang.Double.longBitsToDouble(bits)
 // TODO: the functions below should eventually be intrinsified
 
 inline fun <reified R : Number> Byte.signExtend(): R = when (R::class.java) {
-    java.lang.Byte::class.java -> this.toByte() as R
+    java.lang.Byte::class.java -> this as R
     java.lang.Short::class.java -> this.toShort() as R
     java.lang.Integer::class.java -> this.toInt() as R
     java.lang.Long::class.java -> this.toLong() as R
@@ -55,20 +55,20 @@ inline fun <reified R : Number> Byte.signExtend(): R = when (R::class.java) {
 }
 
 inline fun <reified R : Number> Short.signExtend(): R = when (R::class.java) {
-    java.lang.Short::class.java -> this.toShort() as R
+    java.lang.Short::class.java -> this as R
     java.lang.Integer::class.java -> this.toInt() as R
     java.lang.Long::class.java -> this.toLong() as R
     else -> this.invalidSignExtension()
 }
 
 inline fun <reified R : Number> Int.signExtend(): R = when (R::class.java) {
-    java.lang.Integer::class.java -> this.toInt() as R
+    java.lang.Integer::class.java -> this as R
     java.lang.Long::class.java -> this.toLong() as R
     else -> this.invalidSignExtension()
 }
 
 inline fun <reified R : Number> Long.signExtend(): R = when (R::class.java) {
-    java.lang.Long::class.java -> this.toLong() as R
+    java.lang.Long::class.java -> this as R
     else -> this.invalidSignExtension()
 }
 
@@ -77,20 +77,20 @@ inline fun <reified R : Number> Number.invalidSignExtension(): R {
 }
 
 inline fun <reified R : Number> Byte.narrow(): R = when (R::class.java) {
-    java.lang.Byte::class.java -> this.toByte() as R
+    java.lang.Byte::class.java -> this as R
     else -> this.invalidNarrowing()
 }
 
 inline fun <reified R : Number> Short.narrow(): R = when (R::class.java) {
     java.lang.Byte::class.java -> this.toByte() as R
-    java.lang.Short::class.java -> this.toShort() as R
+    java.lang.Short::class.java -> this as R
     else -> this.invalidNarrowing()
 }
 
 inline fun <reified R : Number> Int.narrow(): R = when (R::class.java) {
     java.lang.Byte::class.java -> this.toByte() as R
     java.lang.Short::class.java -> this.toShort() as R
-    java.lang.Integer::class.java -> this.toInt() as R
+    java.lang.Integer::class.java -> this as R
     else -> this.invalidNarrowing()
 }
 
@@ -98,7 +98,7 @@ inline fun <reified R : Number> Long.narrow(): R = when (R::class.java) {
     java.lang.Byte::class.java -> this.toByte() as R
     java.lang.Short::class.java -> this.toShort() as R
     java.lang.Integer::class.java -> this.toInt() as R
-    java.lang.Long::class.java -> this.toLong() as R
+    java.lang.Long::class.java -> this as R
     else -> this.invalidNarrowing()
 }
